+++ src/docx/index.ts
@@ -3,3 +3,5 @@ export * from "./paragraph";
 export * from "./table";
 export * from "./file";
 export * from "./numbering";
+export * from "./media";
+export * from "./drawing";
