+++ src/docx/media/index.ts
@@ -1,0 +1,2 @@
+export * from "./media";
+export * from "./data";
