+++ src/index.ts
@@ -1,5 +1,3 @@
 export * from "./docx";
 export * from "./export";
-export { Styles } from "./styles";
-export { Media } from "./media";
-export * from "./export";
+export * from "./styles";
